requirements.txt
bidict==0.23.1
blinker==1.9.0
click==8.2.1
colorama==0.4.6
Flask==3.1.2
Flask-SocketIO==5.5.1
h11==0.16.0
itsdangerous==2.2.0
Jinja2==3.1.6
MarkupSafe==3.0.2
python-dotenv==1.1.1
python-engineio==4.12.2
python-socketio==5.13.0
simple-websocket==1.1.0
Werkzeug==3.1.3
wsproto==1.2.0
Flask-SQLAlchemy==3.1.1
#eventlet==0.33.3 not required 
Flask-Dance==7.1.0
requests-oauthlib
google-api-python-client
#gevent not required
# Dependencies for Chatbot
groq
openai
langchain
langchain-core
langchain-community
langchain-groq
# Dependencies for Audio-Analysis (SER/STT)
pyaudio
transformers==4.30.2
soundfile==0.12.1
librosa==0.10.1
torch
numpy==1.24.4
pydub==0.25.1
# Dependencies for Facial-Analysis (FER)
tensorflow==2.16.1
tf-keras==2.16.0
opencv-python==4.9.0.80